Last version I sent to John Nelson in 2013
This code links C++ to Fortran to ObjC.  It implements a nice waterfall
and signal detection for Joe Taylor’s JT65 code.  John did the Qt
implementation and all I did was port it to the new ObjC and make some
adds for automation.  It is here for illustrative purposes mostly.  I
did not modify Joe Taylor’s Fortran (except to make it easier to link
to ObjC).